file reorganization parsers and analyzers
